Create Puzzle_Cube.py
This is the implementation of classes Cube, Move, and Scramble. Cube constructs a cube based on __init__ parameters or defaults to a solved cube. A Cube object prints as numbers that correspond to colors of stickers. Move allows for movement of the cube and the corresponding stickers that rotate accordingly. Scramble provides a random scramble of the cube and gives the user the moves to achieve that scramble